+++ lib/src/styling/app_spacing.dart
@@ -4,6 +4,7 @@ class AppSpacing {
   static const smallerPadding = 4.0;
   static const defaultPadding = 16.0;
   static const zero = 0.0;
+  static const oneLine = 1;
   static const twoLines = 2;
   static const threeLines = 3;
 }